Renamed image_tex to image_mt. Corrected the usage of ^ to | in the new TEX header handling.
