+++ Kuriimu.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.14
+VisualStudioVersion = 15.0.26430.15
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kuriimu", "src\Kuriimu\Kuriimu.csproj", "{7C89FFAF-B638-4577-A1FC-9A783D242428}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -159,8 +159,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "text_bmd", "src\text\text_b
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "image_f3xt", "src\image\image_f3xt\image_f3xt.csproj", "{269D0F0D-AE95-447B-8403-41F28D8DE00C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "image_tex", "src\image\image_tex\image_tex.csproj", "{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "text_srtz", "src\text\text_srtz\text_srtz.csproj", "{9A0C08B2-C1FF-4DE5-B9BB-7267D8FA0EE6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "archive_srtz", "src\archive\archive_srtz\archive_srtz.csproj", "{665782EB-3D6E-4426-B052-5048334D31C0}"
@@ -215,6 +213,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "archive_mt", "src\archive\archive_mt\archive_mt.csproj", "{92407A45-8006-4AA7-95F2-6EED156918A7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "image_mt", "src\image\image_mt\image_mt.csproj", "{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -378,10 +378,6 @@ Global
 		{269D0F0D-AE95-447B-8403-41F28D8DE00C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{269D0F0D-AE95-447B-8403-41F28D8DE00C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{269D0F0D-AE95-447B-8403-41F28D8DE00C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9A0C08B2-C1FF-4DE5-B9BB-7267D8FA0EE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9A0C08B2-C1FF-4DE5-B9BB-7267D8FA0EE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9A0C08B2-C1FF-4DE5-B9BB-7267D8FA0EE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -478,6 +474,10 @@ Global
 		{92407A45-8006-4AA7-95F2-6EED156918A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{92407A45-8006-4AA7-95F2-6EED156918A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{92407A45-8006-4AA7-95F2-6EED156918A7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -515,7 +515,6 @@ Global
 		{482A2B54-F27D-4095-8D91-16B9C03C3403} = {FA7A6EB6-4E0C-4E59-B024-6E8E94A2117C}
 		{CC01BCBF-871A-48AC-9AA9-19507940CFF3} = {B509D23F-0D4E-42F9-87F5-50CF5E982A75}
 		{269D0F0D-AE95-447B-8403-41F28D8DE00C} = {FA7A6EB6-4E0C-4E59-B024-6E8E94A2117C}
-		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4} = {FA7A6EB6-4E0C-4E59-B024-6E8E94A2117C}
 		{9A0C08B2-C1FF-4DE5-B9BB-7267D8FA0EE6} = {B509D23F-0D4E-42F9-87F5-50CF5E982A75}
 		{665782EB-3D6E-4426-B052-5048334D31C0} = {793D1110-C4FD-41A5-9124-6138A5C0B8A7}
 		{22B0F441-5CF7-44CB-BA4F-0F1A25592614} = {FA7A6EB6-4E0C-4E59-B024-6E8E94A2117C}
@@ -540,5 +539,6 @@ Global
 		{2B7589A7-81A9-461F-8BF1-1684093E796A} = {793D1110-C4FD-41A5-9124-6138A5C0B8A7}
 		{B5B621DD-4445-4CBF-95FB-1C3D16A97611} = {B509D23F-0D4E-42F9-87F5-50CF5E982A75}
 		{92407A45-8006-4AA7-95F2-6EED156918A7} = {793D1110-C4FD-41A5-9124-6138A5C0B8A7}
+		{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4} = {FA7A6EB6-4E0C-4E59-B024-6E8E94A2117C}
 	EndGlobalSection
 EndGlobal
