+++ src/image/image_mt/image_mt.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{9E078594-73AE-4EC1-ABA2-BF8F338FCCD4}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>image_tex</RootNamespace>
-    <AssemblyName>image_tex</AssemblyName>
+    <RootNamespace>image_mt</RootNamespace>
+    <AssemblyName>image_mt</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
